breaking change: moved AbstractTarget implementation into AbstractSurreal
